refactor: extract register asset form
