+++ src/pages/assets/register/index.ts
@@ -1,0 +1,2 @@
+export { RegisterAssetForm } from './RegisterAssetForm';
+export { useRegisterAsset } from './useRegisterAsset';
